extract the missing feature dialog and made it accessible anywhere
